### 2016-05-12 #19 @whedon
[REVIEW]: riskscorer - automating risk calculation in clinical practice and research
**Submitting author:** @meyera (<a href="http://orcid.org/0000-0002-6944-2478">Alexander Meyer</a>)
**Repository:** <a href="https://github.com/meyera/riskscorer" target ="_blank">https://github.com/meyera/riskscorer</a>
**Version:** v0.2.0
**Archive:** <a href="http://dx.doi.org/10.5281/zenodo.51342" target ="_blank">http://dx.doi.org/10.5281/zenodo.51342</a>

- [ ] **Archive:** Does the <a target="_blank" href="http://dx.doi.org/10.5281/zenodo.51342">software archive</a> resolve?
- [ ] **Performance:** Have the performance claims of the software been confirmed?
- [ ] **Functionality documentation:** Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?

Compiled paper PDF: [10.21105.joss.00019.pdf](https://github.com/openjournals/joss-reviews/files/261684/10.21105.joss.00019.pdf)
