### 2016-05-12 #19 @Spencerx
I am reviewing this, surge ahead in case of latency
