### 2016-05-13 #19 @Spencerx
So far;
@meyera
Repository is up and running, resolving to mentioned url.
Authors Affiliation and Credit: Checks out ✅
Software License, MIT ✅ check
References are clear enough.

Minor amendments:
You only mispelt "therefore". I'm asserting your audience is in medical or
clinical coding based on the text, this needs an explicit statement in the
abstract in the positive.
So there's a need for a easily accessible risk calculation structure,
engine or kernel to verify the math, with appropriate methods, you probably
need to have formal verification and peer review embedded in the process?
The second paragraph of the abstract is enough to work with for those with
familiarity to clinical or mission critical environments but you need to
offer more detail on the context of extensibility should you want
contributors or such without scaring anyone off yet, what features do you
want to add beyond score methods citing "extensibility"?
By simple programming interface do you mean a command line interface, a
graphic user interface with everyday WIMP or touch based icons, feel free
to elaborate if this will operate exclusively as a headless service or
using any other incarnation you deem appropriate for your evaluation and
operating environments.

Fleshing out the structure before under or over committing to needs will do
enough.
Test data parses fine by dry run reading
https://github.com/meyera/riskscorer/blob/master/README.md
Setting up my cloud for R is taking long
