### 2016-05-13 #19 @arfon
> Authors Affiliation and Credit: Checks out ✅

@Spencerx - feel free to check off items at the top of the issue as you go...
