### 2016-05-22 #19 @meyera
@Spencerx @arfon
Thank you for reviewing my submission. I am not totally sure how the processes work, that's why I just would like to ask: should I start to revise the points mentioned by @Spencerx right now and commit or wait until the review is finished and I get the official decision to revise?
Thank you very much for your effort.
Alex
