### 2016-05-26 #19 @arfon
> should I start to revise the points mentioned by @Spencerx right now and commit or wait until the review is finished and I get the official decision to revise?

@meyera - feel free to make some changes as you go. Just let @Spencerx know that you've made them.
