### 2016-07-26 #19 @meyera
@Spencerx
Please apologize my late answer.
Here are my comments and my revision based on your remarks:

> You only mispelt "therefore". I'm asserting your audience is in medical or
> clinical coding based on the text, this needs an explicit statement in the
> abstract in the positive.

The spelling error is corrected.

> So there's a need for a easily accessible risk calculation structure,
> engine or kernel to verify the math, with appropriate methods, you probably
> need to have formal verification and peer review embedded in the process?

The audicence are scientist and statisticians working in clinical research, The usual risk in clinical medicine are usual simple in structure, but due to the high amount of patients, the manual calculation is a hassle. Even more, there are no tools allowing for batch processing based on a database. The riskscorer package aims to simplify the automatic calculation on demand, for instance as a part of the admission process of a patient, as well as batch processing of patient records.

Verfication for the two implemented scores (EuroScore I and STS score) is accomplished as described below:

_EuroScore I:_
The score is a simple logistic regression model. I have written serveral test cases to compare the calculated score against the a known score for the specific case.

_STS:_
The STS score is more complicated than the EuroScore I. The STS provides a Web-API for riskscore processing, for which, however, no convienent wrappers exists. The risk calculation function is actually a wrapper of the STS risk calculation web service. Test cases were similarily performed as for EuroScore I.

> The second paragraph of the abstract is enough to work with for those with
> familiarity to clinical or mission critical environments but you need to
> offer more detail on the context of extensibility should you want
> contributors or such without scaring anyone off yet, what features do you
> want to add beyond score methods citing "extensibility"?

You are right, extensibility is an exaggeration. In the riskscorer package case extensibility would be just the addition of another calculation function for a dedicated score.
Therefore I deleted the extensibility point in the Paper.md file.

When you refer to 'the abstract', do you mean the Paper.md file?

>By simple programming interface do you mean a command line interface, a
>graphic user interface with everyday WIMP or touch based icons, feel free
>to elaborate if this will operate exclusively as a headless service or
>using any other incarnation you deem appropriate for your evaluation and
>operating environments.

By simple programming interface I mean: each score -> one function. Nothing more. The coding of the various score parameters is flexibly recognized as elaborated in the package's vignette.
