### 2016-08-07 #19 @meyera
@pjotrp @Spencerx @arfon
Thank you for your comment. I agree that one should ask questions whether submitted software to JOSS is a valuable contribution to research.

In regard to the number of lines of code: The _riskscorer_ package consists of several source files.
 - sts.R ==> 863 lines of code
 - test_sts.R ==> 383 lines of code
 - utils.R ==> 581 lines of code
 - test_utils.R ==> 221 lines of code
 - es_II.R ==> 452 lines of code
 - test_esII.R ==> 241 lines of code

It remains unclear to me where you did count 350 lines of code. In general, however, I do think that one should not use the number of lines of code as an indicator for anything.

Regarding the added value of _riskscorer_ to science: calculating riskscores is a tedious task, which is taking a lot of time that could be used for the core research instead. Currently there are no tools that automat risk calculation of the prominent ESII or STS score available. The _riskscorer_ package provides a convenient and time saving way to batch calculate the risk scores. Moreover, the web-service ready design enables integration into the clinical IT infrastructure and workflow and makes it possible to calculate the risk score automatically right at the admission of a patient. Because of the tedious work of the risk calculation in clinical reality these are rarely manually calculated. Automating this process therefor enables the use of powerful tools such as the STS score as a clinical decision support instrument.

We already use the package in our clinical research, papers where this tool was used will be soon ready for submission to clinical journals. Currently the EurValve EU research project looks at the source code to re-implement the code in another language for their own project (http://cordis.europa.eu/project/rcn/199897_en.html).

I hope the value to clinical research that riskscorer gives is now a little clearer.

Yours,
Alex
