### 2016-08-08 #19 @arfon
> I hope the value to clinical research that riskscorer gives is now a little clearer.

👍 thanks for this clear explanation @meyera
