### 2016-09-20 meyera/riskscorer#2 @giacombi
invalid multibyte character in parser at line 734
Dear Meyera,
thank you for the important package you have developed. I'm currently trying to install it, but at the very end it throws an error:

Error in parse(outFile) :
  invalid multibyte character in parser at line 734
ERROR: unable to collate and parse R files for package ‘riskscorer’
* removing ‘/Library/Frameworks/R.framework/Versions/3.3/Resources/library/riskscorer’
Error: Command failed (1)

Thank you for your support.

Giacomo
