### 2016-09-20 #19 @arfon
@Spencerx @meyera - how are we doing on this review?
