### 2016-09-23 meyera/riskscorer#2 @meyera
Thanks for your feedback. I will fix that and update you once it should work.
