### 2016-09-23 meyera/riskscorer#1 @meyera
Thanks for your comment.
I will fix this and update you once it is working.
Greetings,
Alex
