### 2016-10-04 meyera/riskscorer#1 @raffdoc
Hi Alex, sorry If I haven't open new issue I will ask here,
When I feed a data frame to the calc_sts() function as indicated in your vignette it comes back with error like this
`Error: Each variable must be a 1d atomic vector or list.
Problem variables: 'proceduretype'`
What should this mean?
I have looked into the code but this procedure type appears once and I have no idea how to fix it
Many thanks beforehand.
