### 2016-12-12 #19 @arfon
@Spencerx - could you please give us an update on this review?
