### 2017-01-14 #19 @masalmon
I see the repository has two issues that look like bug reports [here](https://github.com/meyera/riskscorer/issues/1) and [here](https://github.com/meyera/riskscorer/issues/2)

What is the status on these issues @meyera ?
