### 2017-01-14 meyera/riskscorer#4 @maelle
Dependencies from the vignette not in DESCRIPTION
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/19)

Not all the packages used in the vignette such as `wakefield` and `threejs` are listed in the dependencies in DESCRIPTION (the should be in suggests). I saw this because of an R CHECK error.
