### 2017-01-14 meyera/riskscorer#5 @masalmon
Continuous integration?
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/19)

The precedent issue would have been identified by continuous integration because the CHECK would have happened with only the dependencies of DESCRIPTION installed. I'd recommend you to set up continuous integration for your repository. It's only a suggestion but I promise it can be very useful and [this post](http://juliasilge.com/blog/Beginners-Guide-to-Travis/) is a beginner-friendly introduction to Travis. I can also help if needed.
