### 2017-01-14 meyera/riskscorer#4 @masalmon
`Cairo` is also such a dependency missing in DESCRIPTION.
