### 2017-01-14 meyera/riskscorer#6 @masalmon
Number of dependencies
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/19)

The package has many dependencies. In my experience, the more dependencies you have, the more issues about installation you will get. Therefore I'd suggest removing some dependencies if possible. E.g. `Cairo` is only needed for having a nice PNG, which can maybe be avoided? I could not find any call to `purrr` (admittedly I didn't search for a long time) except one that is commented.
