### 2017-01-14 meyera/riskscorer#7 @masalmon
R CHECK results
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/19)

```
==> devtools::check(document = FALSE)

Setting env vars ---------------------------------------------------------------
CFLAGS  : -Wall -pedantic
CXXFLAGS: -Wall -pedantic
Building riskscorer ------------------------------------------------------------
"C:/PROGRA~1/R/R-33~1.1/bin/x64/R" --no-site-file --no-environ --no-save  \
  --no-restore --quiet CMD build "C:\Users\Maelle\Documents\riskscorer"  \
  --no-resave-data --no-manual

* checking for file 'C:\Users\Maelle\Documents\riskscorer/DESCRIPTION' ... OK
* preparing 'riskscorer':
* creating vignettes ... OK
* checking for LF line-endings in source and make files
* checking for empty or unneeded directories
* building 'riskscorer_0.2.0.tar.gz'

Setting env vars ---------------------------------------------------------------
_R_CHECK_CRAN_INCOMING_ : FALSE
_R_CHECK_FORCE_SUGGESTS_: FALSE
Checking riskscorer ------------------------------------------------------------
"C:/PROGRA~1/R/R-33~1.1/bin/x64/R" --no-site-file --no-environ --no-save  \
  --no-restore --quiet CMD check  \
  "C:\Users\Maelle\AppData\Local\Temp\RtmpQJCecf/riskscorer_0.2.0.tar.gz"  \
  --as-cran --timings --no-manual

* using log directory 'C:/Users/Maelle/Documents/riskscorer.Rcheck'
* using R version 3.3.1 (2016-06-21)
* using platform: x86_64-w64-mingw32 (64-bit)
* using session charset: ISO8859-1
* using options '--no-manual --as-cran'
* checking for file 'riskscorer/DESCRIPTION' ... OK
* checking extension type ... Package
* this is package 'riskscorer' version '0.2.0'
* checking package namespace information ... OK
* checking package dependencies ... OK
* checking if this is a source package ... OK
* checking if there is a namespace ... OK
* checking for executable files ... OK
* checking for hidden files and directories ... OK
* checking for portable file names ... WARNING
Found the following file with a non-portable file name:
  ext/Abstract UseR2016.md
These are not fully portable file names.
See section 'Package structure' in the 'Writing R Extensions' manual.
* checking whether package 'riskscorer' can be installed ... OK
* checking installed package size ... OK
* checking package directory ... OK
* checking 'build' directory ... OK
* checking DESCRIPTION meta-information ... NOTE
License components which are templates and need '+ file LICENSE':
  MIT
* checking top-level files ... NOTE
File
  LICENSE
is not mentioned in the DESCRIPTION file.
Non-standard files/directories found at top level:
  'ext' 'paper.bib' 'paper.json' 'paper.md'
* checking for left-over files ... OK
* checking index information ... OK
* checking package subdirectories ... OK
* checking R files for non-ASCII characters ... WARNING
Found the following file with non-ASCII characters:
  sts.R
Portable packages must use only ASCII characters in their R code,
except perhaps in comments.
Use \uxxxx escapes for other characters.
* checking R files for syntax errors ... OK
* checking whether the package can be loaded ... OK
* checking whether the package can be loaded with stated dependencies ... OK
* checking whether the package can be unloaded cleanly ... OK
* checking whether the namespace can be loaded with stated dependencies ... OK
* checking whether the namespace can be unloaded cleanly ... OK
* checking loading without being on the library search path ... OK
* checking dependencies in R code ... OK
* checking S3 generic/method consistency ... OK
* checking replacement functions ... OK
* checking foreign function calls ... OK
* checking R code for possible problems ... NOTE
calc_esII: no visible binding for global variable '.'
calc_sts: no visible binding for global variable '.'
calc_sts: no visible binding for global variable 'proceduretype'
calc_sts: no visible binding for global variable 'predmort'
calc_sts: no visible binding for global variable 'predmm'
calc_sts: no visible binding for global variable 'preddeep'
calc_sts: no visible binding for global variable 'predstro'
calc_sts: no visible binding for global variable 'predvent'
calc_sts: no visible binding for global variable 'predreop'
calc_sts: no visible binding for global variable 'predrenf'
calc_sts: no visible binding for global variable 'pred14d'
calc_sts: no visible binding for global variable 'pred6d'
cc_eGFR: no visible binding for global variable '.'
gavs: no visible binding for global variable 'NYHA'
radial_plot: no visible global function definition for 'theme_bw'
radial_plot: no visible global function definition for 'theme'
radial_plot: no visible global function definition for 'element_blank'
radial_plot: no visible global function definition for 'element_rect'
radial_plot: no visible global function definition for 'ggplot'
radial_plot: no visible global function definition for 'xlab'
radial_plot: no visible global function definition for 'ylab'
radial_plot: no visible global function definition for 'coord_equal'
radial_plot: no visible global function definition for 'geom_text'
radial_plot: no visible global function definition for 'aes'
radial_plot: no visible binding for global variable 'x'
radial_plot: no visible binding for global variable 'y'
radial_plot: no visible binding for global variable 'text'
radial_plot: no visible global function definition for
  'scale_x_continuous'
radial_plot: no visible global function definition for
  'scale_y_continuous'
radial_plot: no visible global function definition for 'geom_polygon'
radial_plot: no visible global function definition for 'geom_path'
radial_plot: no visible binding for global variable 'axis.no'
radial_plot: no visible global function definition for 'geom_point'
radial_plot: no visible global function definition for 'labs'
  . NYHA aes axis.no coord_equal element_blank element_rect geom_path
  geom_point geom_polygon geom_text ggplot labs pred14d pred6d preddeep
  predmm predmort predrenf predreop predstro predvent proceduretype
  scale_x_continuous scale_y_continuous text theme theme_bw x xlab y
  ylab
Consider adding
  importFrom("graphics", "text")
to your NAMESPACE file.
* checking for missing documentation entries ... WARNING
Undocumented code objects:
  'calc_esII' 'calc_gavs' 'calc_sts' 'cc_eGFR' 'es_II' 'radial_plot'
All user-level objects in a package should have documentation entries.
See chapter 'Writing R documentation files' in the 'Writing R
Extensions' manual.
 WARNING
'qpdf' is needed for checks on size reduction of PDFs
* checking installed files from 'inst/doc' ... OK
* checking files in 'vignettes' ... OK
* checking examples ... NONE
* checking for unstated dependencies in 'tests' ... OK
* checking tests ...
  Running 'testthat.R' [28s]
 OK
* checking for unstated dependencies in vignettes ... NOTE
'::' or ':::' imports not declared from:
  'threejs' 'tidyr'
'library' or 'require' calls not declared from:
  'ggplot2' 'threejs' 'wakefield'
* checking package vignettes in 'inst/doc' ... OK
* checking re-building of vignette outputs ... OK
* DONE
Status: 4 WARNINGs, 4 NOTEs

See
  'C:/Users/Maelle/Documents/riskscorer.Rcheck/00check.log'
for details.

checking for portable file names ... WARNING
Found the following file with a non-portable file name:
  ext/Abstract UseR2016.md
These are not fully portable file names.
See section 'Package structure' in the 'Writing R Extensions' manual.

checking R files for non-ASCII characters ... WARNING
Found the following file with non-ASCII characters:
  sts.R
Portable packages must use only ASCII characters in their R code,
except perhaps in comments.
Use \uxxxx escapes for other characters.

checking for missing documentation entries ... WARNING
Undocumented code objects:
  'calc_esII' 'calc_gavs' 'calc_sts' 'cc_eGFR' 'es_II' 'radial_plot'
All user-level objects in a package should have documentation entries.
See chapter 'Writing R documentation files' in the 'Writing R
Extensions' manual.
 WARNING
'qpdf' is needed for checks on size reduction of PDFs

checking DESCRIPTION meta-information ... NOTE
License components which are templates and need '+ file LICENSE':
  MIT

checking top-level files ... NOTE
File
  LICENSE
is not mentioned in the DESCRIPTION file.
Non-standard files/directories found at top level:
  'ext' 'paper.bib' 'paper.json' 'paper.md'

checking R code for possible problems ... NOTE
calc_esII: no visible binding for global variable '.'
calc_sts: no visible binding for global variable '.'
calc_sts: no visible binding for global variable 'proceduretype'
calc_sts: no visible binding for global variable 'predmort'
calc_sts: no visible binding for global variable 'predmm'
calc_sts: no visible binding for global variable 'preddeep'
calc_sts: no visible binding for global variable 'predstro'
calc_sts: no visible binding for global variable 'predvent'
calc_sts: no visible binding for global variable 'predreop'
... 26 lines ...
radial_plot: no visible global function definition for 'geom_point'
radial_plot: no visible global function definition for 'labs'
  . NYHA aes axis.no coord_equal element_blank element_rect geom_path
  geom_point geom_polygon geom_text ggplot labs pred14d pred6d preddeep
  predmm predmort predrenf predreop predstro predvent proceduretype
  scale_x_continuous scale_y_continuous text theme theme_bw x xlab y
  ylab
Consider adding
  importFrom("graphics", "text")
to your NAMESPACE file.

checking for unstated dependencies in vignettes ... NOTE
'::' or ':::' imports not declared from:
  'threejs' 'tidyr'
'library' or 'require' calls not declared from:
  'ggplot2' 'threejs' 'wakefield'
R CMD check results
0 errors | 3 warnings | 4 notes

R CMD check succeeded
```

There are a few warnings (ignore the one about `qpdf`) and notes. It is important to solve them, even more so if you intend to submit this package to CRAN which I'd recommend doing.
