### 2017-01-14 meyera/riskscorer#8 @masalmon
Link to other packages
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/19)

I have two remarks here:

* You mention it is the first package to offer such functionality. However, if there are packages that have functionalities that might complement this one, e.g. from the [CRAN Task View clinical trials](https://cran.r-project.org/web/views/ClinicalTrials.html) could you mention them, e.g. in the README or vignette?

* If you end up submitting the package to CRAN (which as I said in another issue I'd recommend because it makes your package easier to find, trust and install), maybe it could be added to a relevant CRAN task view?
