### 2017-01-14 meyera/riskscorer#9 @masalmon
README
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/19)

* The README has an unfinished sentence starting with "Eventually".

* When is the future release containing the two other algorithms planned? The answer to this question won't affect my review conclusions, it's just that I think that users would be curious about this.
