### 2017-01-14 meyera/riskscorer#10 @masalmon
Test coverage
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/19)

First I need to mention that it's already enough for JOSS + really cool that this package has tests.

I ran `covr::package_coverage()` and got

```r
> covr::package_coverage()
riskscorer Coverage: 65.94%
R/gavs.R: 0.00%
R/utils.R: 35.29%
R/sts.R: 77.33%
R/esII.R: 80.51%
```

Maybe test coverage could be increased a little?

To visualize the lines of code that don't get tested you could:

* use the `covr::shine` function

* enable Travis to send coverage reports to e.g. `codecov.io` as explained [here](https://github.com/jimhester/covr) if you set up continuous integration (which was only a suggestion). The reason I'd recommend doing this is that you can add a badge with the test coverage to your README, and this way users can see that your package has tests so it helps trust.
