### 2017-01-14 meyera/riskscorer#11 @masalmon
Documentation
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/19)

In the doc of `cal_sts` you need to escape the "%" in e.g. ""60%" or "45 %"" because otherwise the arguments after this one are not rendered properly in the .Rd file.
