### 2017-01-14 meyera/riskscorer#12 @masalmon
License
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/19)

Your package includes a reproduction of http://www.euroscore.org/ What is the license of the original algorithm?
