### 2017-01-14 meyera/riskscorer#13 @masalmon
Terms of use of STS web calculator
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/19)

I don't understand the fact that the STS web calculator can't be used according to the terms of use, but it is in your package? Could you clarify this point?

Moreover if it is the case this point needs to be mentioned on top of the function documentation because the function documentation is very long.
