### 2017-01-14 meyera/riskscorer#14 @masalmon
Empty .R
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/19)

I don't understand why there is [a nearly empty R script?](https://github.com/meyera/riskscorer/blob/master/R/esI.R)
