### 2017-01-14 meyera/riskscorer#15 @masalmon
Function(s) in development
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/19)

Regarding the `gavs` function which is in development, I'd recommend not exporting it or exporting it in a development branch of the package.
