### 2017-01-14 meyera/riskscorer#16 @masalmon
Vignette
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/19)

* On Figure 1 there's no legend for the colors which make the graph hard to read.

* Also maybe an example with only 2 variables and therefore in 2D would be easier to read as a first example? This is only a suggestion.

* It'd be important to have a comment for each figure. The comment would explain how the risk varies with the variable(s) and how this is expected by using the score value.

* In the review thread you write "I have written serveral test cases to compare the calculated score against the a known score for the specific case." I'd encourage you to add this info in the documentation of the package because as an user it's nice to know without having to look into the tests file.

* The vignette could include the formula / principle of the risk score.

* "Terms of use of STS web calculator does not permit" -> "Terms of use of STS web calculator do not permit"

* Typo "Therefor"
