### 2017-01-14 meyera/riskscorer#6 @masalmon
I think the dependency on `plumber` is unnecessary. It's nice to mention its possible use, but you can have the code describing it in not evaluated chunks/doc and avoid the dependency.
