### 2017-01-14 meyera/riskscorer#17 @masalmon
Typo in es_II doc
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/19)

"FUnction for calculation without fuzzy parameter matching"
