### 2017-01-14 meyera/riskscorer#18 @masalmon
Too many exported functions?
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/19)

I wonder why `parse_vd_regurg` and `parseVesselsDisease` are exported: they're not used by the user, are they?
