### 2017-01-14 #19 @masalmon
Congratulations on creating this package. I can see how it can be useful to have the risk scorer as R functions instead of web interfaces, and your package could have a larger list of such scorers in the future. Here is my review, and I've opened several issues in the software repository itself.

## Reviewer questions
### General checks
- [x ] **Repository:** Is the source code for this software available at the <a target="_blank" href="https://github.com/meyera/riskscorer">repository url</a>?
- [ x] **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
Yes but it gets an R CHECK error because apparently the license file needs to be mentioned in the DESCRIPTION.

- [ x] **Version:** Does the release version given match the GitHub release (v0.2.0)?
- [x ] **Archive:** Does the <a target="_blank" href="http://dx.doi.org/10.5281/zenodo.51342">software archive</a> resolve?
### Functionality
The reason I don't check this box yet is the fact that not all dependencies used in the vignette are listed in the DESCRIPTION so if one tried to install the package & build the vignette at the same time there'd be an error.

- [ ] **Performance:** Have the performance claims of the software been confirmed?

The reasons I leave this unchecked are 1) the examples in the vignette are not commented at all so it's hard to see if this is the expected behaviour of the functions 2) as mentioned in the issue about the vignette, it'd be nice to mention the existence of tests comparing values obtained with the package against known values.

### Documentation
- [x ] **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
- [x ] **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
- [ x] **Functionality documentation:** Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?

There's a special character issue in one Rd file at least though so please check the lists of arguments.

- [x ] **Automated tests:** Are there automated tests or manual steps described so that the function of the software can be verified?

No, and this would be really important, e.g. if someone wishes to add a score calculator to your package.

### Software paper

Compiled paper PDF: [10.21105.joss.00019.pdf](https://github.com/openjournals/joss-reviews/files/261684/10.21105.joss.00019.pdf)
- [ x] **Authors:** Does the `paper.md` file include a list of authors with their affiliations?
- [ x] **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
- [ x] **References:** Do all archival references that should have a DOI list one (e.g. papers, datasets, software)?
