### 2017-01-20 #19 @arfon
@meyera - please reply here when you've had a chance to address/respond to @masalmon's review.
