### 2017-01-26 #19 @meyera
@masalmon thank you very much for the fast review. I will start working on it and inform you once everything's done.
