### 2017-01-26 #19 @masalmon
Cool, feel free to tag me here or in the repo if I can help!
