+++ R/acc_sts_tavr_score.R
@@ -1,0 +1,454 @@
+#' Calculates the ES II score
+#'
+#' FUnction for calculation without fuzzy parameter matching
+#'
+#' @param NYHA character vector with the following values: NULL or 'I', 'II', 'III', 'IV'
+#' @param IDDM numeric 0 or 1
+#' @param age numeric age
+#' @param female numeric 0 or 1
+#' @param ECA numeric 0 or 1
+#' @param CPD numeric 0 or 1
+#' @param nm_mob numeric 0 or 1
+#' @param redo numeric 0 or 1
+#' @param renal_dysfunc character vector with the following values: NULL or 'gt_85', '50_85', 'lt_50', 'dialysis'
+#' @param active_endo numeric 0 or 1
+#' @param critical numeric 0 or 1
+#' @param lv_func character vector with the following values: NULL or 'gt_50', '31_50', '21_30', 'lt_20'
+#' @param recent_mi numeric 0 or 1
+#' @param sPAP character vector with the following values: NULL or 'lt_31', '31_55', 'gt_55'
+#' @param urgency character vector with the following values: NULL or 'elective' or 'urgent', 'emergency', 'salvage'
+#' @param proc_weight character vector with the following values: NULL or 'isolated-CABG', 'single-non-CABG', '2-procedures', '3-procedures'
+#' @param thoracic_aorta numeric 0 or 1
+#'
+#' @return predicted probability of mortality
+#' @export
+es_II <- function(age,
+                  female,
+                  NYHA = NULL,
+                  CCS4 = NULL,
+                  IDDM = NULL,
+                  ECA = NULL,
+                  CPD = NULL,
+                  nm_mob = NULL,
+                  redo = NULL,
+                  renal_dysfunc = NULL,
+                  active_endo = NULL,
+                  critical = NULL,
+                  lv_func = NULL,
+                  recent_mi = NULL,
+                  sPAP = NULL,
+                  urgency = NULL,
+                  proc_weight = NULL,
+                  thoracic_aorta = NULL
+                  ) {
+
+  const <- -5.324537
+  logOdds <- const
+
+  age_coef <- 0.0285181
+  if (age <=60) {
+    logOdds <- logOdds + 1*age_coef
+  } else if (age >60) {
+    logOdds <- logOdds + ((age - 59) * age_coef)
+  }
+
+  if (!is.null(NYHA)) {
+    if (NYHA == 'II') {
+      logOdds <- logOdds + 0.1070545
+    } else if (NYHA == 'III') {
+      logOdds <- logOdds + 0.2958358
+    } else if (NYHA == 'IV') {
+      logOdds <- logOdds + 0.5597929
+    }
+  }
+
+  if (!is.null(female)) {
+    logOdds <- logOdds + (female * 0.2196434)
+  }
+
+  if (!is.null(renal_dysfunc)) {
+    if (renal_dysfunc == 'gt_85') {
+      logOdds <- logOdds + 0
+    } else if (renal_dysfunc == '50_85') {
+      logOdds <- logOdds + 0.303553
+    } else if (renal_dysfunc == 'lt_50') {
+      logOdds <- logOdds + 0.8592256
+    } else if (renal_dysfunc == 'dialysis') {
+      logOdds <- logOdds + 0.6421508
+    }
+  }
+
+  if (!is.null(ECA)) {
+    logOdds <- logOdds + (ECA * 0.5360268)
+  }
+
+  if (!is.null(nm_mob)) {
+    logOdds <- logOdds + (nm_mob * 0.2407818)
+  }
+
+  if (!is.null(redo)) {
+    logOdds <- logOdds + (redo * 1.118599)
[…]
