+++ R/AccTavi.R
@@ -1,0 +1,194 @@
+# to do:
+# - default values in für alle Funktionen festlegen. 
+# NA or NULL? bei eigentlicher predRisk-Formel NULL sinnvoll. bei calcAcuity NA sinnvoller, da Argumente keine "optional arguments", sondern logicals
+# - Testen
+# - testthat error messages einheitlich. "or", "," einheitlich mit Alex
+# - zulässige Argumente einheitlich festlegen mit Alex und einheitliche Beschreibung bei @param X
+
+
+#' @param age numeric between 18 and 100
+#' @param sex Valid values: "f", "female", "F", "Female", "w" or "m", "male", "M", "Male"
+#' @param race "w", "white", "other", "Other", NA -> white race or other than black race
+#'             "b", "black" -> black or African American
+#' @param access Access site. Valid values: NA, "femoral", "Femoral" or "Non-femoral", "non-femoral"
+#' @param crea Valid values: 0.1 - 150 (mg/dL = default unit) or 8.8 - 13260 (µmol/L Unit must be changed in param unit)
+#' @param unit Default value: mg/dL, if needed, change to "SI" (µmol/L)
+#' @param dialysis      Valid values: "yes", "y", "Yes", "Y", 1, "no", "n", "No", "N", 0, NA
+#' @param nyha      Valid values: "yes", "y", "Yes", "Y", 1, "no", "n", "No", "N", 0, NA
+#' @param scld      Valid values: "yes", "y", "Yes", "Y", 1, "no", "n", "No", "N", 0, NA
+#' @param procStatus    Valid values: NA or "elective", "urgent", "emergency", "salvage"
+#' @param pcArrest      Valid values: "yes", "y", "Yes", "Y", 1, "no", "n", "No", "N", 0, NA
+#' @param ppInotr   Valid values: "yes", "y", "Yes", "Y", 1, "no", "n", "No", "N", 0, NA
+#' @param pcShock       Valid values: "yes", "y", "Yes", "Y", 1, "no", "n", "No", "N", 0, NA
+#' @param ppMechan  Valid values: "yes", "y", "Yes", "Y", 1, "no", "n", "No", "N", 0, NA
+
+
+
+# parseIfValid:
+#   argument x: vector (e.g dialysis, access...)
+#   argument vector1: vector of valid values for x (e.g c("yes", "y", "Yes", "Y", 1))
+#   argument vector0: vector of valid values for x (e.g c("no", "n", "No", "N", 0)))
+#   returns parsedValue: 1 if x in vector1, 0 if x in vector0. If x is neither in vector1 
+#   nor in vector0, an error message occurs.
+# e.g.:
+# parseIfValid("y", c("Y", "y", 1), c("N", 0))
+
+parseIfValid <- function(x, vector1, vector0){
+  if (x %in% vector1){
+    parsedValue <- 1
+  } else if (x %in% vector0){
+    parsedValue <- 0
+  } else {
+    stop ("Invalid ", substitute(x), ". Valid values are: ", toString(vector0), " or ", toString(vector1))
+  }
+  return(parsedValue)
+}
+
+
+# calcGFR: 
+#   returns a list with GFR, ageValue and dialysisValue. All three are needed for risk calculation in calcRisk.
+# e.g.:
+# calcGFR(age = 33, "f", 150, unit = "SI")
+
+calcGFR <- function(age, sex, crea, race = "white", unit = "US", dialysis = NA){
+  boolYes <- c("yes", "y", "Yes", "Y", 1)
+  boolNo <- c("no", "n", "No", "N", 0, NA)
+  
+  if ( age < 18 | age > 100 | !is.numeric(age)) {
+    stop ('age is no numeric between 18 and 100')
[…]
