+++ tests/Testing_AccTavi.R
@@ -1,0 +1,266 @@
+source("Functions3.R")
+library(testthat)
+
+# toDo:
+#   testthat error messages einheitlich. "or", "," einheitlich mit Alex
+
+
+# Test if error messages occur in case of invalid values
+validValues <- list(age = 50, sex = "f", crea = 1.5, access = "femoral", nyha = 0,
+                    scld = 0, procStatus = "elective", pcArrest = 0, pcShock = 0,
+                    ppInotr = 0, ppMechan = 0, race = "white", dialysis = 0,
+                    unit = "US")
+# e.g. do.call(calcRisk, validValues)
+
+context("ACC Internal Calculation - Error messages")
+
+test_that("Error message, if age is no numeric between 18 and 100", {
+  a <- validValues
+  a$age <- 17
+  expect_error(do.call(calcRisk, a), "age is no numeric between 18 and 100")
+  a$age <- 101
+  expect_error(do.call(calcRisk, a), "age is no numeric between 18 and 100")
+  a$age <- "19"
+  expect_error(do.call(calcRisk, a), "age is no numeric between 18 and 100")
+})
+test_that("Error message, if sex is invalid", {
+  b <- validValues
+  b$sex <- "fe"
+  expect_error(do.call(calcRisk, b), 'Invalid sex. Valid values are: f, female, F, Female, w or m, male, M, Male')
+  b$sex <- "ma"
+  expect_error(do.call(calcRisk, b), 'Invalid sex. Valid values are: f, female, F, Female, w or m, male, M, Male')
+  b$sex <- 0
+  expect_error(do.call(calcRisk, b), 'Invalid sex. Valid values are: f, female, F, Female, w or m, male, M, Male')
+})
+test_that("Error message, if access is invalid", {
+  c <- validValues
+  c$access <- "fem"
+  expect_error(do.call(calcRisk, c), "Invalid access. Valid values are: femoral, Femoral, NA or non-femoral, Non-femoral")
+  c$access <- 2
+  expect_error(do.call(calcRisk, c), "Invalid access. Valid values are: femoral, Femoral, NA or non-femoral, Non-femoral")
+  c$access <- "normal"
+  expect_error(do.call(calcRisk, c), "Invalid access. Valid values are: femoral, Femoral, NA or non-femoral, Non-femoral")
+})
+test_that("Error message, if nyha is invalid", {
+  e <- validValues
+  e$nyha <- "a"
+  expect_error(do.call(calcRisk, e), "Invalid nyha. Valid values are: no, n, No, N, 0, NA or yes, y, Yes, Y, 1" )
+  e$nyha <- 2
+  expect_error(do.call(calcRisk, e), "Invalid nyha. Valid values are: no, n, No, N, 0, NA or yes, y, Yes, Y, 1" )
+  e$nyha <- 10
+  expect_error(do.call(calcRisk, e), "Invalid nyha. Valid values are: no, n, No, N, 0, NA or yes, y, Yes, Y, 1" )
+})
+test_that("Error message, if scld is invalid", {
+  f <- validValues
+  f$scld <- "a"
+  expect_error(do.call(calcRisk, f), "Invalid scld. Valid values are: no, n, No, N, 0, NA or yes, y, Yes, Y, 1" )
+  f$scld <- 2
+  expect_error(do.call(calcRisk, f), "Invalid scld. Valid values are: no, n, No, N, 0, NA or yes, y, Yes, Y, 1" )
+  f$scld <- 10
[…]
