### 2017-07-10 #19 @maelle
@meyera cc @arfon as mentioned I can clarify my feedback and help if needed
